Создал controller.py, в котором собраны функции, для реализации веб-приложений
