+++ start.py
@@ -1,0 +1,4 @@
+if __name__ == "__main__":
+    from controller import *
+
+    app.run(debug=True)
